integration-main/integration-main/requirements-combined.txt
paho-mqtt==1.6.1
toml==0.10.2
requests==2.31.0
sqlalchemy>=2.0.0
types-sqlalchemy
alembic==1.13.1
bcrypt==4.1.2
flask==3.0.2
flask-login
flask-sqlalchemy
flask-socketio
flask-session
python-dotenv
flask-wtf
psycopg2-binary==2.9.9
werkzeug==3.0.1
black==24.1.1
Flask-Login==0.6.3
mypy==1.8.0
mypy-extensions==1.0.0
pylint==3.0.3
pytest==8.0.0
types-requests==2.31.0.20240125
types-toml==0.10.8.7
geopy
